docs: Add guide for cluster with Docker Hub
The current setup guide is complicated and requires setup of Harbor and minikube. We assume the average developer looking at Connaisseur has a test cluster available and is fine with testing on Docker Hub for a start. Then this guide is much shorter and focusses on the Connaisseur aspects of image integrity.