ci/docs: Release concept for Connaisseur image and release check on new tagged commit on master branch (#44)
Previously we had no real concept of how to release images. We now considered automatic releases and automatic image signing, but decided against it. This MR documents the decision and adds a pipeline that checks whether manual release was successful upon a new tag being created on the master branch. Moreover, it is ensured that the tag matches the versions defined and the values.yml is now source of truth for the version in the Makefile.