ci: Improve tag detection during release (#78)
Previously we used the GITHUB_REF variable set during the execution of of Github Actions to get to the current tag, which failed if tagging was done on prior to the push/on a different branch. Now we use the information contained in the git repository that is checked out anyway. Additionally the release pipeline previously also ran for any new commit on master (without a new tag being present) and now runs only for tags in potentially any branch (which is fine, as the release pipeline does no privileged action).