fix: missing delegation files
in the special case, that delegations haven been added to an imgae repository, but no signatures were created using said delegations, the delegations appear in the targets JSON, but the actual delegation JSON files don't exist yet. Previously connaisseur assumed them to be present if they are in the targets JSON. This was changed to use the targets in the targets JSON file, instead of the missing delegation JSON files.