feat: implementation of simple alerting interface (#70)
Co-authored-by: Philipp Belitz <philipp.belitz@securesystems.de>
Co-authored-by: annekebr <44376590+annekebr@users.noreply.github.com>

This allows to configure external entities to receive notifications either if connaisseur admits or if it rejects a deployment. There are templates for Slack, OpsGenie as well as Keybase, but it is built to allow custom extensions easily. Notifications are implemented as simple POST requests with a configurable JSON payload and extendable headers.

Co-authored-by: Philipp Belitz <philipp.belitz@securesystems.de>