fix: ensure backwards compatibility upgrading to 1.5.0
Upgrading to 1.5.0 without pulling the newest version of the helm templates will result in a Configuration Error as there are configuration files missing that are needed for the alerting which came with 1.5.0. This fix ensures that one can update to the new image without failures even without repulling the connaisseur repo. The added feature will, however, not be functional without the new helm templates.