hotfix: fixed image format validation
The image validation assumed that the registry/domain name of an image contains a dot (.) as domain usually do. However this doesn't work for images coming from private registries like 'private-registry:3000'. Making the dot optional also doesn't fix the problem since for image like 'path/to/image:tag', 'path' would be recognized as a domain name. Thus the validation has been change so that domains that do not contain a dot, necessarily need to specify a port in order for them to be recognized as domains.

fix #114