fix: health probe fails after some time (#166)
After installing connaisseur the health probe will fail after some time though it only returns true it runs health checks on validators that will not influence the health status but be logged if unhealthy. In case the health checks for the validators take too long, the pod health probe timeouts and the pod is considered unhealthy. Validator health checks are now separated from the pod health.