docs/fix: fix get-root utility by (#183)
Previously, building the get-root utility failed due to a new requirement of the cryptography package, that now requires a rust compiler on alpine-based distributions. Since that bloats image and build time, this commit switches to a debian base image. Also, the utility was out of sync with the changes to the notary interface, which is also rectified by this commit. Lastly, it fixes a faulty reference in the docs

Fix 184