fix: webhook api version
Fixed an error, where the patch version of Kubernetes in some implementations (microk8s) could look different from what was expected. The regex was generalized so it should properly match now. Additionally for Kubernetes v1.16 and below the 'v1beta1' api version of the webhook is used (previously v1.15 and below).