fix: digest confusion at high load
it seems that as of v2.5.0 digest confusion can occur, i.e. a referenced image gets the digest of another image assigned and patched accordingly which causes deployments potentially falsly allowed, but failing due to broken request patching. the root issue appears to be unexpectatly related to updating a class variable inside concurrent futures / asyncio. the class variable is removed and inputs are passed into the concurrent tasks, results returned and then used to update the resuts

fixes #583