fix: exclude k8s CRDs from parent resource check
Per default, Connaisseur checks the parent resource of a k8s object as part of automatic child approval. This fails when creating a workload object from custom resource definitions (CRDs) as the API group and thus the structure is not known. This is fixed by excluding the parent check for unknown API groups

fixes #582