build: Remove transitive dependency pinning
In light of #620 we pinned a transitive dependency to fix incompatibility introduced by a breaking change in said transitive dependency. While being a reasonable quick fix at the time, it imposes undue limits on our dependency. Thus this commit removes the pinning of the transitive dependency and should they again break (the transitive dependency took notice and the dependency took steps to mitigate), we may need to reeavluate the dependency as a whole instead of pinning dependencies that are not ours to pin.