ci: ignore CVE-2022-28948
CVE-2022-28948 (https://avd.aquasec.com/nvd/2022/cve-2022-28948/) affects Go-Yaml v3 package which is a dependency of cosign. The vulnerable package appears to only affect cosign functionality that is not exposed in the Connaisseur context. Furthermore, cosign itself is called as Subprocess cmd which simply returns failed validation in case an attacker manages to induce a crash by exploiting the vulnerability.