test: Test workload readiness
This commit introduces checks that not only are workloads in our workload tests deployed, but that they actually become ready at some point. This is related to #514, which reports that we successfully deployed a parent resource, which never became ready as its child resources were invalid. Introducing a readiness check ensures that any resources we check are actually functional, in turn preventing regression alike to #514.

Fix #516