ci: acquire sa token
For the "other-ns" integration test a service account (SA) was created to make request to the kube api as this SA. Since k8s v1.22 SAs do not automatically generate their own access tokens as secrets anymore, but create them ad-hoc. The integration test was changed accordingly to this changed bahavior.