fix: revert frozenlist changes
There was a problem with the frozenlist package, which didn't had its packages added to wheels (https://github.com/aio-libs/frozenlist/issues/342). The packages are now added, so the fix for this can be reverted, except for the pytest job which is the only one building on a debian image (faster runtime). Here the yarl and multidict packages still have the same problem as frozenlist had. Waiting fo a fix here.