feat: update to cosign 2.0 and switch to asyncio
With cosign 2.0 some defaults where changed, breaking the normal connaisseur workflow. E.g. verification with the tranperency log is now on by default. Introduced a flag for image pattern to enable/disable this feature. Also the validation of multiple signature for cosign was refactored by switching from futures to asyncio.