fix: less verbose logging for cosign
With the functionality for cosign validators to validate multiple signatures with different keys and meeting certain thresholds, the logging output for these can be confusing. It is expected that certain keys do not match a signature and that an error is logged, even though the overall verification is successfull. To reduce confusion here and to make it clear the verification was actually sucessfull, the error logging was set to debug instead of info.

fixes #977