fix: skip replica to zero
When updating a deployment form image A to B a new replicaset is created for image B and the old replicaset is set to 0 replicas. Previously Connaisseur tried to validated the previous replicasets update request, which may fail because image A no longer has a valid signature. This validation is now skipped.

fixes #1224