fix: do not redundantly authenticate calls to notary
Currently, for each TUF role the first request is simply to get the auth endpoint, the second is getting the actual auth token and only the third fetches the actual data. For four roles that makes 12 requests in total. Now, the auth token should be fetched beforehand (2 calls) resulting in 6 calls in total.