test: Resolve sporadic integration test failures
Before this commit the integration testing of ephemeral containers failed sporadically in the sense that more admissions than anticipated happened. This PR awaits the base pod that is to be debugged by the ephemeral container as the unreadiness of the base pod led to another admission request being sent once the base pod got ready.