sec: Prevent ReDoS during delegation validation
This commit fixes a regular expression denial of service vulnerability that was present during the schema validation of the targets.json TUF file in a Notary validation. An adversary with the ability to alter Notary responses, in particular an evil Notary server, could have provided Connaisseur with a malformed targets.json, which would have led to catastrophic (exponential) backtracking during regex matching. Exploitability was contingent on the presence of a notaryv1 validator in the Connaisseur configuration, which is the default as both DockerHub's and Connaisseur's Notary root keys are present in the default values.yaml.