fix: no exceptions on automatic child approval
When automatic child approval is active and can't find the parent resource, even though there is one, the validation should continue, as if there were no parent resources in the first place. This prevents some deployments from failing, where the parent resource was deleted and now the child can never be validated again.