build: unpin ca certificates
The version for the ca-certificates package, installed during an intermediate build step from an alpine image, is being unpinned. Reason for that is that alpine only ever provides a single version for their packages and once they update the ca-certificates packages, the Connaisseur build automatically fails, since the previous used version is no longer available.