fix: return empty patch type if there is no patch
When Connaisseur was done with the verification and returned an empty patch, the patch type was still set to JSONPatch. Kubernetes handled this regardless, but other tools like ArgoCD did not and threw an error. Connaisseur will now neither set a patch nor patch type and leave those fields empty, so it's compatible with other tools.