test: fix flakey redis-cert test
The redis cert test tested that after upgrading a redis with a custom certificate the caching mechanism still worked. Problem was, the redis instance might have been ready on Kubernetes side, but internally not ready to receive connections. Thus the caching was tested and failed, before the newly upgraded redis was ready. This has been fixed with a wating mechanism.