test: fixed failing workload test
The workload test tried to test a StatefulSet workload, which in case tried to claim a volume. To do so it tried to startup another pod which Connaisseur would block, because of the deny all default policy. The workload test was now decoupled from the preconfig test so that it gets their own custom policy in which the extra pod is allowed. Also some refactoring on the waiting of readiness for workloads.